stacks: use "--allowed-groups" to restrict types stacks can depend on
This is a draft towards #899, allowing the Stack Manager to be run
with a --allowed-groups parameter that limits the GVK groups that a
Stack may depend on.

```
$CROSSPLANE stack manage \
  --allowed-groups cache.crossplane.io,compute.crossplane.io,...
```

One weaknesses of this approach is that new stacks can not be used as
dependencies.

`compute.gcp.crossplane.io` would not be a viable dependency with the
example above. One can imagine wanting to depend on newly installed
Stacks to layer the Crossplane experience.  This may not be a problem,
if only core claim types are acceptable, but core claim types should
perhaps not be given special treatment (per #1154).

Signed-off-by: Marques Johansson <marques@upbound.io>